+++ SearchableNavbarCommon/SearchableNavbarMarginFactory.cs
@@ -7,6 +7,7 @@ using Microsoft.VisualStudio.ComponentModelHost;
 using Microsoft.VisualStudio.Editor;
 using Microsoft.VisualStudio.Shell;
 using Microsoft.VisualStudio.Shell.Interop;
+using Microsoft.VisualStudio.Text;
 using Microsoft.VisualStudio.Text.Editor;
 using Microsoft.VisualStudio.Utilities;
 
@@ -23,12 +24,17 @@ namespace SearchableNavbar
 #pragma warning disable CS0649
         [Import(typeof(SVsServiceProvider))]
         private IServiceProvider ServiceProvider;
+        [Import]
+        internal ITextDocumentFactoryService TextDocumentFactoryService;
 #pragma warning restore CS0649
 
         public IWpfTextViewMargin CreateMargin(IWpfTextViewHost wpfTextViewHost, IWpfTextViewMargin marginContainer)
         {
             ThreadHelper.ThrowIfNotOnUIThread();
-            return new SearchableNavbarMargin(wpfTextViewHost.TextView, ServiceProvider.GetService(typeof(DTE)) as DTE2, ServiceProvider.GetService(typeof(SVsImageService)) as IVsImageService2);
+            return new SearchableNavbarMargin(wpfTextViewHost.TextView, 
+                ServiceProvider.GetService(typeof(DTE)) as DTE2, 
+                ServiceProvider.GetService(typeof(SVsImageService)) as IVsImageService2,
+                TextDocumentFactoryService);
         }
     }
 }