+++ CHANGELOG.md
@@ -7,3 +7,8 @@
 ### Version 1.12 (31/05/2023)
 
 - Hotfix for lots of redundant processes being created and slowing down Visual Studio
+
+### Version 1.13
+
+- Searchbar now stays hidden if there's nothing to show
+- A document is parsed again if it's saved through Visual Studio after being previously modified